+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Breaking Changes
   - `const {MockGuild} = require('chaos-core').test.discordMocks`
   - `const {MockCommand} = require('chaos-core').test.chaosMocks`
 - Removed Mockery from chaos mocks
+- Removed `services` and `commands` from config
 
 Minor Features
 --------------